Testing new approach
To be frank, I do not like the way this site looks at all with the
awkward perspectives. I feel it hinders basic use in addition to being
unpleasing aesthetically.

I think I have found a better direction using a little JavaScript to
allow the perspective origin to scroll with the user.